databricks_dashboard/requirements.txt
# SFDC Consumption & Forecast Dashboard
databricks-sdk
databricks-sql-connector
streamlit
pandas
openpyxl
python-dateutil
